+++ terraform-azure-modules/examples/complete-landing-zone/main.tf
@@ -33,6 +33,53 @@ module "log_analytics" {
   tags                = module.tags.tags
   name                = module.naming.log_analytics
 
-  # Forwarding to a separate central workspace (e.g. Sentinel) can be enabled later by setting this object.
   diagnostics_settings = null
+}
+
+module "vnet" {
+  source = "../../modules/networking/virtual-network"
+
+  resource_group_name = module.rg.name
+  location            = "uksouth"
+  tags                = module.tags.tags
+  name                = module.naming.virtual_network
+  address_space       = ["10.60.0.0/16"]
+
+  diagnostics_settings = {
+    log_analytics_workspace_id = module.log_analytics.id
+  }
+}
+
+module "nsg" {
+  source = "../../modules/networking/network-security-group"
+
+  resource_group_name = module.rg.name
+  location            = "uksouth"
+  tags                = module.tags.tags
+  name                = module.naming.nsg
+  security_rules      = []
+
+  diagnostics_settings = {
+    log_analytics_workspace_id = module.log_analytics.id
+  }
+}
+
+module "subnet" {
+  source = "../../modules/networking/subnet"
+
+  resource_group_name  = module.rg.name
+  location             = "uksouth"
+  tags                 = module.tags.tags
+  name                 = module.naming.subnet
+  virtual_network_name = module.vnet.name
+  address_prefixes     = ["10.60.1.0/24"]
+
+  diagnostics_settings = {
+    log_analytics_workspace_id = module.log_analytics.id
+  }
+}
+
+resource "azurerm_subnet_network_security_group_association" "main" {
+  subnet_id                 = module.subnet.id
+  network_security_group_id = module.nsg.id
 }