+++ terraform-azure-modules/modules/networking/route-table/outputs.tf
@@ -1,1 +1,14 @@
-# Placeholder outputs will be defined when this module is implemented.
+output "id" {
+  description = "The route table resource ID."
+  value       = azurerm_route_table.this.id
+}
+
+output "name" {
+  description = "The route table name."
+  value       = azurerm_route_table.this.name
+}
+
+output "route_table" {
+  description = "The route table resource object."
+  value       = azurerm_route_table.this
+}