+++ terraform-azure-modules/modules/networking/subnet/locals.tf
@@ -1,3 +1,3 @@
 locals {
-  # Reserved for name construction and tag merging when implemented.
+  subnet_name = var.name
 }