+++ terraform-azure-modules/modules/networking/subnet/outputs.tf
@@ -1,1 +1,14 @@
-# Placeholder outputs will be defined when this module is implemented.
+output "id" {
+  description = "The subnet ID."
+  value       = azurerm_subnet.this.id
+}
+
+output "name" {
+  description = "The subnet name."
+  value       = azurerm_subnet.this.name
+}
+
+output "subnet" {
+  description = "The subnet resource object."
+  value       = azurerm_subnet.this
+}