+++ terraform-azure-modules/modules/networking/virtual-network/outputs.tf
@@ -1,1 +1,14 @@
-# Placeholder outputs will be defined when this module is implemented.
+output "id" {
+  description = "The Azure resource ID of the virtual network."
+  value       = azurerm_virtual_network.this.id
+}
+
+output "name" {
+  description = "The name of the virtual network."
+  value       = azurerm_virtual_network.this.name
+}
+
+output "virtual_network" {
+  description = "The virtual network resource object."
+  value       = azurerm_virtual_network.this
+}